adding reorient_image
