+++ ivadomed/image.py
@@ -1,0 +1,26 @@
+import nibabel as nib
+import numpy as np
+from ivadomed.loader import utils as imed_loader_utils
+
+
+def reorient_image(arr, slice_axis, nib_ref, nib_ref_canonical):
+    """Reorient an image to match a reference image orientation.
+
+    It reorients a array to a given orientation and convert it to a nibabel object using the reference nibabel header.
+
+    Args:
+        arr (ndarray): Input array, array to re orient.
+        slice_axis (int): Indicates the axis used for the 2D slice extraction: Sagittal: 0, Coronal: 1, Axial: 2.
+        nib_ref (nibabel): Reference nibabel object, whose header is used.
+        nib_ref_canonical (nibabel): `nib_ref` that has been reoriented to canonical orientation (RAS).
+    """
+    # Orient image in RAS according to slice axis
+    arr_ras = imed_loader_utils.orient_img_ras(arr, slice_axis)
+
+    # https://gitship.com/neuroscience/nibabel/blob/master/nibabel/orientations.py
+    ref_orientation = nib.orientations.io_orientation(nib_ref.affine)
+    ras_orientation = nib.orientations.io_orientation(nib_ref_canonical.affine)
+    # Return the orientation that transforms from ras to ref_orientation
+    trans_orient = nib.orientations.ornt_transform(ras_orientation, ref_orientation)
+    # apply transformation
+    return nib.orientations.apply_orientation(arr_ras, trans_orient)
